Move GetUserDataByPID to utility
