+++ database/wiiu/get_user_friend_list.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/PretendoNetwork/friends/database"
 	"github.com/PretendoNetwork/friends/globals"
+	"github.com/PretendoNetwork/friends/utility"
 	"github.com/PretendoNetwork/nex-go"
 	friends_wiiu_types "github.com/PretendoNetwork/nex-protocols-go/friends-wiiu/types"
 )
@@ -55,13 +56,7 @@ func GetUserFriendList(pid uint32) ([]*friends_wiiu_types.FriendInfo, error) {
 		} else {
 			// Offline
 
-			userData, err := globals.GetUserData(friendPID)
-			if err != nil {
-				globals.Logger.Critical(err.Error())
-				continue
-			}
-
-			userInfo, err := GetUserInfoByPNIDData(userData)
+			userInfo, err := utility.GetUserInfoByPID(friendPID)
 			if err != nil {
 				globals.Logger.Critical(err.Error())
 				continue
